Extract tweet processing to cranberry microservice
