+++ docker-compose.yml
@@ -13,6 +13,7 @@ services:
       DATAPUNNET_ADMIN_PASSWORD: ${DATAPUNNET_ADMIN_PASSWORD}
       POSTGRES_HOST: ${POSTGRES_HOST}
       REDIS_HOST: ${REDIS_HOST}
+      REDIS_DATABASE: ${REDIS_DATABASE}
       REDIS_PASSWORD: ${REDIS_PASSWORD}
       RABBITMQ_HOST: ${RABBITMQ_HOST}
       RABBITMQ_USERNAME: ${RABBITMQ_USERNAME}
@@ -20,11 +21,5 @@ services:
       MNIST_ENDPOINT: ${MNIST_ENDPOINT}
       LDA_ENDPOINT: ${LDA_ENDPOINT}
       TIMESERIES_ENDPOINT: ${TIMESERIES_ENDPOINT}
-      TWITTER_CONSUMER_KEY: ${TWITTER_CONSUMER_KEY}
-      TWITTER_CONSUMER_SECRET: ${TWITTER_CONSUMER_SECRET}
-      TWITTER_ACCESS_TOKEN: ${TWITTER_ACCESS_TOKEN}
-      TWITTER_ACCESS_SECRET: ${TWITTER_ACCESS_SECRET}
-      TWITTER_FEED: ${TWITTER_FEED}
-      TWITTER_TIMEOUT: ${TWITTER_TIMEOUT}
     ports:
       - 80:80
