+++ test/stub/bayberry/service/redis.ex
@@ -1,8 +1,11 @@
 defmodule Stub.Bayberry.Service.Redis do
   alias Supervisor.Spec
 
-  def command(_conn, _command) do
-    {:ok, nil}
+  def command(_conn, command) do
+    case command do
+      ["get", "twitter:track"] -> {:ok, "#hashtag"}
+      _ -> {:ok, nil}
+    end
   end
 
   def json(file) do
